requirements.txt
absl-py==0.10.0
astroid==2.4.2
astunparse==1.6.3
attrs==20.2.0
cachetools==4.1.1
certifi==2020.6.20
cffi==1.14.3
chardet==3.0.4
cycler==0.10.0
Cython==0.29.21
dataclasses==0.6
dill==0.3.2
dm-tree==0.1.5
flatbuffers==1.12
future==0.18.2
gast==0.3.3
gin-config==0.3.0
google-api-core==1.22.4
google-api-python-client==1.12.3
google-auth==1.22.1
google-auth-httplib2==0.0.4
google-auth-oauthlib==0.4.1
google-cloud-bigquery==2.1.0
google-cloud-core==1.4.3
google-crc32c==1.0.0
google-pasta==0.2.0
google-resumable-media==1.1.0
googleapis-common-protos==1.52.0
grpcio==1.32.0
h5py==2.10.0
httplib2==0.18.1
idna==2.10
importlib-metadata==2.0.0
importlib-resources==3.0.0
isort==5.6.4
joblib==0.17.0
kaggle==1.5.8
Keras-Preprocessing==1.1.2
kiwisolver==1.2.0
lazy-object-proxy==1.4.3
Markdown==3.3.1
matplotlib==3.3.2
mccabe==0.6.1
numpy==1.19.2
oauth2client==4.1.3
oauthlib==3.1.0
opencv-python-headless==4.4.0.44
opt-einsum==3.3.0
pandas==1.1.3
Pillow==7.2.0
portpicker==1.3.1
promise==2.3
proto-plus==1.10.1
protobuf==3.13.0
psutil==5.7.2
py-cpuinfo==7.0.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycocotools==2.0.2
pycparser==2.20
pylint==2.6.0
pyparsing==2.4.7
python-dateutil==2.8.1
python-slugify==4.0.1
pytz==2020.1
PyYAML==5.3.1
requests==2.24.0
requests-oauthlib==1.3.0
rsa==4.6
scikit-learn==0.23.2
scipy==1.5.2
sentencepiece==0.1.91
seqeval==1.1.0
six==1.15.0
slugify==0.0.1
tb-nightly==2.4.0a20201012
tensorboard-plugin-wit==1.7.0
tensorflow-addons==0.11.2
tensorflow-datasets==4.0.1
tensorflow-hub==0.9.0
tensorflow-metadata==0.24.0
tensorflow-model-optimization==0.5.0
termcolor==1.1.0
text-unidecode==1.3
tf-estimator-nightly==2.4.0.dev2020101301
tf-models-nightly==2.3.0.dev20201013
tf-nightly==2.4.0.dev20201012
tf-slim==1.1.0
threadpoolctl==2.1.0
toml==0.10.1
tqdm==4.50.2
typed-ast==1.4.1
typeguard==2.9.1
typing-extensions==3.7.4.3
uritemplate==3.0.1
urllib3==1.24.3
Werkzeug==1.0.1
wrapt==1.12.1
yapf==0.30.0
zipp==3.3.0
